requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf>=5.28.0
redis==5.1.1
kafka-python==2.0.2
requests==2.32.3
python-dotenv